Updating the validation/parser for Custom Flag tracking
We've had some reports that some values our customers can enter for a
custom flag entry, which are valid per documentation, fail validation
when being used in the SDK. This PR attempts to change that issue by
making the definition and validation of Variable entries more
consistent. However, it needs to be verified whether this is correct,
and whether anything needs to change on the back-end as well.